Dependencies added:
The Rpacks directory and CRISPRDesignVenv are added which contain all
python and R dependencies to run the pipeline.

The Azimuth code from GitHub had to be refactored from python2 to
python3, and the models used to infer guide potency needed to be
retrained, as the version of Scikit Learn used by the group is very old
(Python 2.0).

README.md documentation was added as well.